requirements.txt
# Core Deep Learning
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.9
albumentations==1.3.1

# Data Processing
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Model Training & Monitoring
tensorboard==2.15.1
wandb==0.16.1

# Metrics & Evaluation
scikit-image==0.22.0

# Web Interface (for demo app)
streamlit==1.29.0
gradio==4.12.0

# Video Processing
imageio==2.33.1
imageio-ffmpeg==0.4.9

# 3D Visualization (for avatar)
pyrender==0.1.45
trimesh==4.0.8

# Text-to-Sign
nltk==3.8.1
spacy==3.7.2

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
ipywidgets==8.1.1

# Type hints and code quality
typing-extensions==4.9.0

# For reproducibility
python-dateutil==2.8.2
